requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.13.0
tzlocal==5.3.1
urllib3==2.3.0
Flask==3.1.0
asyncpg==22.3.1